requirements.txt
# Core API
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9      # required for FastAPI file uploads

# ML / Inference
torch==2.3.0
torchvision==0.18.0
timm==1.0.3                  # backbone pretrained weights

# Document Layout Detection
# NOTE: detectron2 must be installed AFTER torch — see install instructions below
layoutparser==0.3.4

# Image processing
Pillow==10.3.0
opencv-python-headless==4.9.0.80   # headless = no GUI deps (important for Docker)
numpy==1.26.4

# Validation & config
pydantic==2.7.1
pydantic-settings==2.3.0     # for environment variable config

# Testing
pytest==8.2.2
httpx==0.27.0                # async test client for FastAPI

# Utilities
python-dotenv==1.0.1
loguru==0.7.2                # better logging than stdlib
